fix(repo): migrate verified file moves to recoverable transaction
